Move match endpoints to config
