+++ zen/should_block_request.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/AikidoSec/firewall-go/internal/agent"
 	"github.com/AikidoSec/firewall-go/internal/agent/config"
-	"github.com/AikidoSec/firewall-go/internal/http"
 	"github.com/AikidoSec/firewall-go/internal/log"
 	"github.com/AikidoSec/firewall-go/internal/request"
 )
@@ -25,9 +24,8 @@ func ShouldBlockRequest(ctx context.Context) *BlockResponse {
 		return &BlockResponse{"blocked", "user", nil}
 	}
 	// rate-limiting :
-	matches := http.MatchEndpoints(
-		http.RouteMetadata{URL: reqCtx.URL, Method: reqCtx.Method, Route: reqCtx.Route},
-		config.GetEndpoints(),
+	matches := config.MatchEndpoints(
+		config.RouteMetadata{URL: reqCtx.URL, Method: reqCtx.Method, Route: reqCtx.Route},
 	)
 
 	for _, endpoint := range matches {
